Move commands to separate file
